fully refactored/normalized the code base; split classes into their own module pair of .h/.cpp files, added CMake configuration and generation to incorporate more effective versioning and production times
